% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.r
\name{available_seq_prop}
\alias{available_seq_prop}
\title{Internal function to check which properties users have chosen}
\usage{
available_seq_prop(df_sequence)
}
\arguments{
\item{df_sequence}{dataframe}
}
\value{
df_sequence
}
\description{
Internal function to check which properties users have chosen
}
