% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{rnalab_scatterplot}
\alias{rnalab_scatterplot}
\title{Plot Scatterplots for input data}
\usage{
rnalab_scatterplot(data, x, y, fit)
}
\arguments{
\item{data}{Dataset to use for ggplot2 plots}

\item{x, y}{The x and y variables to be used in the aes() mapping for ggplot scatterplots}

\item{fit}{A boolean indicating whether a linear regression fit should be added to the ggplot scatterplot}
}
\value{
A grid graphical object consisting of a ggplot2 jitter plot with or without a linear regression fit.
}
\description{
Use ggplot2 to generate Scatterplots for input data.
Allows the option to plot a linear regression fit.
}
\examples{
# Plot a scatter plot without linear regression added (using data from the dnaseqs data set)
rnalab_scatterplot(dnaseqs, x = 'length', y = 'yield', fit = FALSE)
# Plot a scatter plot with linear regression added (using data from the dnaseqs data set)
rnalab_scatterplot(dnaseqs, x = 'length', y = 'yield', fit = TRUE)
}
